adding emp onboarding apis

+++ src/main/java/com/ralien/erp_system/employee/dao/DeptRepository.java
@@ -3,5 +3,5 @@ package com.ralien.erp_system.employee.dao;
 import com.ralien.erp_system.employee.entities.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeptRepository extends JpaRepository<Department, Short> {
+public interface DeptRepository extends JpaRepository<Department, Integer> {
 }
